Bloom effect using Shaders
When you stare at a bright light for too long your vision
goes kind of blurry with spots. Don't try it!
This is the effect we implement in this code. Note the shaders
are defined and written in GLSL and are in the Media/Shaders folder
To understand them you will need to look at a GLSL tutorial -
http://www.lighthouse3d.com/tutorials/glsl-tutorial/