BugFix: Fixed issue with PauseState disconnection
Jack noticed this bug in the in class test of this codebase. Thanks
Jack! The problem was that the pausestate was stopping packets from
being passed to the MultiplayerGameState which was underneath it on
the stack. Needed to set PauseState::Update to allow updates
through in Network mode. Also need to change StateStack to call
OnActivate and OnDestroy when changing states. This allows pausing
input from a state and then allowing it again.